+++ scripts/irccat.py
@@ -55,10 +55,10 @@ def main():
     irc.logging.setup(args)
     target = args.target
 
-    client = irc.client.Manifold()
+    reactor = irc.client.Reactor()
     try:
-        c = client.server().connect(args.server, args.port, args.nickname)
-    except irc.client.ServerConnectionError:
+        c = reactor.server().connect(args.server, args.port, args.nickname)
+    except irc.reactor.ServerConnectionError:
         print(sys.exc_info()[1])
         raise SystemExit(1)
 
@@ -66,7 +66,7 @@ def main():
     c.add_global_handler("join", on_join)
     c.add_global_handler("disconnect", on_disconnect)
 
-    client.process_forever()
+    reactor.process_forever()
 
 if __name__ == '__main__':
     main()