+++ scripts/ssl-cat.py
@@ -54,15 +54,15 @@ def main():
     target = args.target
 
     ssl_factory = irc.connection.Factory(wrapper=ssl.wrap_socket)
-    client = irc.client.Manifold()
+    reactor = irc.client.Reactor()
     try:
-        c = client.server().connect(
+        c = reactor.server().connect(
             args.server,
             args.port,
             args.nickname,
             connect_factory=ssl_factory,
         )
-    except irc.client.ServerConnectionError:
+    except irc.reactor.ServerConnectionError:
         print(sys.exc_info()[1])
         raise SystemExit(1)
 
@@ -70,7 +70,7 @@ def main():
     c.add_global_handler("join", on_join)
     c.add_global_handler("disconnect", on_disconnect)
 
-    client.process_forever()
+    reactor.process_forever()
 
 if __name__ == '__main__':
     main()